Minor code refactors in `app.data` package
